refactor: notice query 로직 분리
