+++ apps/host/src/entities/notice/api/notice.ts
@@ -1,9 +1,6 @@
-import { queryOptions } from '@tanstack/react-query';
-
 import { del, get, post, put } from '@amp/apis';
 
 import { END_POINT } from '@shared/constants/end-point';
-import { ORGANIZERS_QUERY_KEY } from '@shared/constants/query-key';
 import type {
   CreateNoticeBody,
   CreateNoticeResponse,
@@ -74,15 +71,3 @@ export const deleteNotice = (noticeId: number) =>
 
 export const getNoticeDetail = (noticeId: number) =>
   get<NoticeDetail>(END_POINT.GET_NOTICE_DETAIL(noticeId));
-
-export const NOTICE_QUERY_OPTIONS = {
-  DETAIL: (noticeId: number | null) => {
-    const normalizedNoticeId = noticeId ?? Number.NaN;
-
-    return queryOptions({
-      queryKey: ORGANIZERS_QUERY_KEY.NOTICE_DETAIL(normalizedNoticeId),
-      queryFn: () => getNoticeDetail(normalizedNoticeId),
-      enabled: Number.isFinite(normalizedNoticeId),
-    });
-  },
-} as const;
